Look at the key type to determine the expected algorithm.
It's not mandatory to have an `alg` claim in the JWKS; we can
derive the information we need from the `kty` (key type) claim.
The java library used by IAM does not advertise `alg`.

There's potentially more work to do here if we want to support
additional RSA signing algorithms.