Add cmake option to turn on/off -Werror (warnings are errors)
-Werror is unconditonally on before this change, which is now considered
poor practice -- it means that the library may refuse to build on
arbirtray compilers or systems that add new warnings.  In particular,
systems with openssl 3 generate an ocean of deprecation warnings.

This change adds a cmake option to turn -Werror back on with

cmake -DSCITOKENS_WARNINGS_ARE_ERRORS:bool=true ...