First attempt at an async refactoring of the SciTokens library.
This splits operations that may block on network access into a "start"
and "continue" portion.  The start portion returns a status object
which may be queried for information about whether the operation is
finished and information about file descriptors that are useful for
select.

Completely untested but it compiles and is roughly the top-to-bottom
outline of what the async code will look like.