Shorten timeout for public key download
While the public key download can fail and have the library fall
back on the cached copy, this fallback can take such a long time
that the application code (such as the HTCondor-CE) invoking the
library times out its current operation and still failing.

So, without this, the fallback succeeds but everything else fails
due to timeouts.

For now, we have arbitrarily set the timeout to 30s for the case
where the pubkey has expired and 4s for the update check.