Add support for API-configurable cache home
There has been desire to create a method by which users can configure
where to store the key cache without setting the env var `$XDG_CACHE_HOME`
or relying on the user to have a home dir. This is especially relevant
in cases where the library is embedded in a daemon (e.g. XRootD or HTCondor)
that may have a preferencial key cache location. This commit adds a familiar
`config_set_str` and `config_get_str` API for passing a new cache home to
the library through the use of the `keycache.cache_home` key.

As with `config_set_int` and `config_get_int`, these new functions can be
easily updated in the future as the project develops and new
configuration keys are desired. This would entail adding a new check for
the desired key and the logic to handle any values associated with that key.