Add duplicate config APIs with prefixed names
Unfortunately, the choice of names for the config_set_blah() family of APIs is
quite generic and poses the potential to cause C symbol collisions when SciTokens
is used by large projects that dynamically load lots of libraries (a la HTCondor).
A reasonable plan to address this, since those APIs are in use and changing their
names would be API-breaking, seems to be adding a second set of prefixed APIs, with
the intention of deprecating the old functions at the next major release.